+++ src/sixtypical/model.py
@@ -19,13 +19,13 @@ TYPE_BYTE = Type('byte')
 TYPE_BYTE_TABLE = Type('byte table')
 
 
-class InteractionConstrainedType(Type):
+class ExecutableType(Type):
     """Used for routines and vectors."""
     def __init__(self, name, inputs=None, outputs=None, trashes=None):
         self.name = name
-        self.inputs = inputs or []
-        self.outputs = outputs or []
-        self.trashes = trashes or []
+        self.inputs = inputs or set()
+        self.outputs = outputs or set()
+        self.trashes = trashes or set()
 
     def __repr__(self):
         return 'RoutineType(%r, inputs=%r, outputs=%r, trashes=%r)' % (
@@ -44,13 +44,13 @@ class InteractionConstrainedType(Type):
         return hash(self.name) ^ hash(self.inputs) ^ hash(self.outputs) ^ hash(self.trashes)
 
 
-class RoutineType(InteractionConstrainedType):
+class RoutineType(ExecutableType):
     """This memory location contains the code for a routine."""
     def __init__(self, **kwargs):
         super(RoutineType, self).__init__('routine', **kwargs)
 
 
-class VectorType(InteractionConstrainedType):
+class VectorType(ExecutableType):
     """This memory location contains the address of a routine."""
     def __init__(self, **kwargs):
         super(VectorType, self).__init__('vector', **kwargs)